layout of the list of courses completed  pagination

+++ database/migrations/2019_05_04_222641_create_reviews_table.php
@@ -19,7 +19,7 @@ class CreateReviewsTable extends Migration
             $table->foreign('course_id')->references('id')->on('courses');
             $table->unsignedInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
-            $table->float('rating',2);
+            $table->integer('rating');
             $table->text('comment')->nullable();
             $table->timestamps();
         });
